Simplify the signature of `query_exec` and `insert_query_job`.
The only information needed for a BQ query is (1) a query string and (2) a
project ID to bill the query to. This change simplifies `query_exec` to
only *require* these two fields, and packages the rest into
`destination_table` and `default_dataset`.

This is a decidedly breaking change for both `query_exec` and
`insert_query_job`; we probably want to bump a version number before merging
this. This also updates all references (I could find) to use the new
`query_exec` signature.

Also update all documentation.