Fix CSV quote encoding.
BigQuery expects strings to be quoted in the "usual" manner, meaning that
quotes in string fields should be doubled, with the field itself in quotes.

This accomplishes this by tweaking arguments to `encodeString` and
`write.table`, and (unfortunately) quotes all string fields as a side-effect.
Adds one simple test of CSV encoding.

Fixes #30.