Port all the underlying HTTP code to use httr2.
This commit refactors `bigrquery`'s internals to use `httr2` instead of
`httr`. I don't have a strong motivation for this change, other than the
general belief that we should be trying to move packages to `httr2` over
time, and because it made `bigrquery`'s internals a bit more modular.

One of the more visible consequences is that the OAuth tokens returned
by `bq_token()` can no longer take the form of an `httr` request object;
now they are simple bearer tokens instead. This should make it much
easier to implement viewer-based credentials in the future.

However, it might break downstream packages that rely on the existing
`bq_token()` return value.

I believe that most of these changes should be covered by the existing
test suite.

Signed-off-by: Aaron Jacobs <aaron.jacobs@posit.co>